+++ tests/fixtures/common/command.json
