--- tests/fixtures/readarr/command.json
@@ -1,47 +1,0 @@
-[
-  {
-    "id": 0,
-    "name": "string",
-    "commandName": "string",
-    "message": "string",
-    "body": {
-      "sendUpdatesToClient": true,
-      "updateScheduledTask": true,
-      "completionMessage": "string",
-      "requiresDiskAccess": true,
-      "isExclusive": true,
-      "isTypeExclusive": true,
-      "name": "string",
-      "lastExecutionTime": "2021-12-12T17:40:12.167Z",
-      "lastStartTime": "2021-12-12T17:40:12.167Z",
-      "trigger": "unspecified",
-      "suppressMessages": true,
-      "clientUserAgent": "string"
-    },
-    "priority": "normal",
-    "status": "queued",
-    "queued": "2021-12-12T17:40:12.167Z",
-    "started": "2021-12-12T17:40:12.167Z",
-    "ended": "2021-12-12T17:40:12.167Z",
-    "duration": {
-      "ticks": 0,
-      "days": 0,
-      "hours": 0,
-      "milliseconds": 0,
-      "minutes": 0,
-      "seconds": 0,
-      "totalDays": 0,
-      "totalHours": 0,
-      "totalMilliseconds": 0,
-      "totalMinutes": 0,
-      "totalSeconds": 0
-    },
-    "exception": "string",
-    "trigger": "unspecified",
-    "clientUserAgent": "string",
-    "stateChangeTime": "2021-12-12T17:40:12.167Z",
-    "sendUpdatesToClient": true,
-    "updateScheduledTask": true,
-    "lastExecutionTime": "2021-12-12T17:40:12.167Z"
-  }
-]
