--- tests/fixtures/sonarr/command.json
@@ -1,8 +1,0 @@
-{
-  "name": "RescanSeries",
-  "startedOn": "0001-01-01T00:00:00Z",
-  "stateChangeTime": "2014-02-05T05:09:09.2366139Z",
-  "sendUpdatesToClient": true,
-  "state": "pending",
-  "id": 24
-}
